requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==23.1.0
click==8.1.3
colorama==0.4.6
Django==3.2.18
django-htmx==1.14.0
django-taggit==3.1.0
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==17.6.0
flake8==6.0.0
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.1
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.2.2
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
